+++ lib/cookbook.recipe.php
@@ -47,6 +47,8 @@
 		}
 		
 		public function delete($matches) {
+			if (!$this->app->loggedId()) { return $this->app->error403(); }
+
 			$id = $matches[1];
 			$this->db->deleteRecipe($id);
 			HTTP::redirect("/");
